% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts1314.R
\docType{data}
\name{counts1314}
\alias{counts1314}
\title{Expression counts for the 2013-2014 Rainout Shelter drought experiment}
\format{A data frame with 17770 rows and 51 variables:
\describe{
  \item{rownames}{the names of each gene, Panicum virgatum annotation (v2.1)}
  \item{colnames}{the names of each library, described in the
  accompanying info1314 dataset}
}}
\usage{
counts1314
}
\description{
A dataset containing the counts of 17770 genes
and 51 individuals, grown under rainout shelters at either th USDA facility
in Temple, TX or the Ladybird Johnson Wildflower Center, Austin TX.
Libraries were sequenced following the 3' tag-SEQ protocol in the Juenger
Laboratory at the University of Texas, Austin. Counts were called via a custom
BWA - HTSeq pipeline, curtosy of Scott Schwartz, 1-Feb 2016
}

