% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countscyl.R
\docType{data}
\name{countscyl}
\alias{countscyl}
\title{Expression counts for the cylinder drought experiment}
\format{A data frame with 24865 rows and 62 variables:
\describe{
  \item{rownames}{the names of each gene, Panicum virgatum annotation (v2.1)}
  \item{colnames}{the names of each library, described in the
  accompanying infocyl dataset}
}}
\usage{
countscyl
}
\description{
A dataset containing the counts of 24865 genes
and 62 individuals, grown in cylinders at the Pickle
Research Center, Austin, TX in 2011. Libraries were sequenced
following the 3' tag-SEQ protocol in the Juenger Laboratory at the
University of Texas, Austin. Counts were called via a custom
BWA - HTSeq pipeline, curtosy of Scott Schwartz, 1-Feb 2016
}

