% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infocyl.R
\docType{data}
\name{infocyl}
\alias{infocyl}
\title{Experimental design data for the 2011 Pickle Cylinder drought experiment}
\format{A data frame with 62 rows and 17 variables:
\describe{
  \item{pot}{A unique identifier for the plant, used as the repeated measure}
  \item{geno}{The Switchgrass genotype: only AP13}
  \item{id}{The library preparation ID. Corresponds to column names in countscyl}
  \item{trt}{The experimental treatment. Wet or Dry}
  \item{psi}{Midday leaf water potential, measured at the time of tissue collection}
  \item{RWC}{Relative water content of leaf, measured at the time of tissue collection}
  \item{LDMC}{Leaf dry mass content, measured at the time of tissue collection}
  \item{spad}{Leaf dry mass content, measured at the time of tissue collection}
  \item{psi.delta}{Leaf dry mass content, measured at the time of tissue collection}
  \item{HHMMSS}{Time of measurment / time of tissue collection}
  \item{FTime}{..., measured at the time of tissue collection}
  \item{Photo}{Photosynthetic rate, measured at the time of tissue collection}
  \item{Cond}{Stomatal Conductance, measured at the time of tissue collection}
  \item{Ci}{Internal Leaf CO2 concentration, measured at the time of tissue collection}
  \item{time_trt}{Character vector corresponding to all time*treatment combinations,
  Used for contrasts}
  \item{time_trt.mat}{Numeric vector corresponding to all time*treatment combinations,
  Used for contrasts}
}}
\usage{
infocyl
}
\description{
A dataset containing the physiology and experimental design for
62 individuals, grown in cylinders at the Pickle Research Campus of UT Austin.
More detailed information regarding the timing and days of sample collection
can be found in the supplementary material of Lovell, Shakirov et al.
}

